fastapi-demo/requirements.txt
fastapi-0.116.1
uvicorn-0.35.0
sqlalchemy-2.0.41
asyncpg-0.30.0
pydantic_settings-2.10.1
alembic-1.16.4
python_dotenv-1.1.1
psycopg2_binary-2.9.10